Add files via upload
In this blog you will be looking at the statistical analysis and visuals of the indian premier league dataset.The dataset is fetched from website named “Tableau Public”.  The visualisations in this article are mainly influenced by attributes like batsmen runs count, bowling records and player performance. So, let's get started with working with data and making useful visualizations.